login, join components정리 완료
