+++ lib/_core/constants/http.dart
@@ -1,0 +1,14 @@
+
+import 'package:dio/dio.dart';
+import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+
+final dio = Dio(
+  BaseOptions(
+    //자기 IP넣기
+    baseUrl: "http://192.168.0.59:8080",
+    contentType: "application/json; charset=utf-8",
+    validateStatus: (status) => true,
+  ),
+);
+
+const secureStorage = FlutterSecureStorage();
