+++ lib/holder/MyPage/creator/components/creator_my_page_header.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
+import '../../../../_core/constants/theme.dart';
 import '../../_components/my_page_custom_button.dart';
 import '../../_components/my_page_order_mileage.dart';
-import '../../../../theme.dart';
 
 class CreatorMyPageheader extends StatelessWidget {
   @override
