+++ lib/holder/MyPage/creator/components/creator_view_page_header.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import '../../../../_core/constants/theme.dart';
 import 'package:flutter/widgets.dart';
 
-import '../../../../theme.dart';
 
 class CreatorViewPageheader extends StatelessWidget {
   @override
