+++ lib/main.dart
@@ -1,9 +1,7 @@
-import 'package:final_project_team02/theme.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-
+import '_core/constants/theme.dart';
 import 'holder/buy/_address_tab.dart';
-import 'holder/main/home_page.dart';
 import 'holder/main_holder.dart';
 
 void main() {
@@ -24,7 +22,6 @@ class _MyAppState extends State<MyApp> {
       debugShowCheckedModeBanner: false,
       theme: theme(),
       home: MainHolder(),
-
     );
   }
 }
