+++ pubspec.yaml
@@ -41,6 +41,13 @@ dependencies:
   intl: ^0.18.1
   ionicons: ^0.2.2
   remedi_kopo: ^0.0.2
+  # 라이브러리 버전 문제가 발생하면 pub dev에서 최신 버전으로 업그레이드하세요
+  flutter_svg: ^2.0.6
+  validators: ^3.0.0
+  dio: ^5.2.0 # 서버와 통신하기 위해 필요한 라이브러리 입니다.
+  flutter_riverpod: ^2.3.6 # 상태관리 Riverpod 라이브러리 입니다.
+  logger: ^1.3.0 # 콘솔창에서 결과물을 쉽게 확인할 수 있도록 하는 Log 라이브러리입니다.
+  flutter_secure_storage: ^8.0.0 # 어플리케이션 Secure Storage를 쉽게 사용할 수 있도록 도와주는 라이브러리입니다.
 
 dev_dependencies:
   flutter_test:
